**** IMPORTANT ****
I have unset the TCL_LIBRARY and TK_LIBRARY variables in tcltk.c.  The
tclUnixInit fn's in the tcl distribtions check these files for places
to look for the standard library tcl files.  This change stops this
at other sites, but NOT here.

The problem still exists here because tcl compiles in the location of
TCL_LIBRARY and TK_LIBRARY in making the binary library files.  So
only if sites have their tcl/tk installations installed in the same
place as us will this be a problem.  It is really important that the
files in rat/tcl and rat/tk match those in the library we are linking
against.  I believe the problem people have been reporting about
menu's not working, etc, was caused by this sourcing problem - it was
with tcl7.6.