WARNING DO NOT UPDATE IF YOU HAVE A FREEBSD / LINUX / IRIX MACHINE
In order to add support for the osprey sound card (SunVideo Plus) I
ammodifying tthe suond card interfaces code.  it now looks like
most of the rest of the code (table of function pointers).

Solaris works with sunaudio and win32 works, but have not got around
to the rest yet.

should be ready by the end of monday.