Holy Tomali.  Here is a bug and a half...
The null audio device code did not take account of samples remaining
in "device" buffer when returning how many samples where available.
As a result it always read a little slow, 25% slower!

So what I was trying to figure out was why when I run rat of a
reflector it sounds okay, but when I run two on the same machine using
the null device rat as the transmitter, the rat with the proper device
would recalculate the playout every few packets and break up the audio
since it's buffer kept running dry.

And a lot of time was wasted looking at the receiver trying to spot
the problem.  Ouch.  This probably exists in the FreeBSD PC speaker
driver also, will check now.