- A redundancy encoder.  Seems to work, but let's not advertise it yet.
I have just wasted 36 hours on a really nasty bug / feature.  So
here's the story if you are interested.

There was a memory "leak" in the redundancy encoder, except it wasn't
leaking as pointers still referenced it.  Basically, coded units come
in from a playout buffer in the transmitter.  As they come in they get
detached and put on another playout buffer in the redundancy encoder
until we have enough units to make a layer of redundancy and churn out
a packet.  As layers go out they strip the units off of the redundancy
playout buffer, leaving empty node in the playout ready to cleaned up.
Except I made a mistake and was clearing up the history in the playout
buffer from the transmitter which has no side effects since the
transmitter also audits this buffer.  So memory was being invested in
empty nodes in the playout buffer of the redundancy encoder.  It's not
"leaking" per se as it is all still referenced.  And the rate of
investment is pretty slow, a few bytes per node, and 50 nodes per
second.

The problem arises with block_alloc.  For those who don't know the
purpose of block_alloc is to be able to recycle memory without
invoking malloc since this invokes a system call and is expensive on
some hardware.  It does this by rounding up the size of the block
requested to a convenient boundary and maintains a list of the blocks
of that size.  When a request for a block comes in, block_alloc checks
the block list for an appropriate block and returns one from the list
if there's one there, or allocates another one.  One the block gets
released with block_alloc it goes onto the list of blocks.  There is
some subtlety to how this works in practice, but it is irrelevent
here.

When in debugging mode block_alloc stores the names of the files and
the line number of where the memory allocation requests come using the
macro's __FILE__ and __LINE__.  The problem is that memory was being
used up in one place and allocated somewhere else as two of the
rounded data structure sizes rounded to the same size.  Whilst nodes
in the playout buffer where not being released, it looked like another
part of the redundancy encoder was using them up since that part of
the code always allocated units of that size.  A large amount of time
then went into locating the "missing" data of that type by adding
copious counts of blocks allocated and freed in various stages of the
redundancy and channel types code and there wasn't any going awry.

The bug is in block_alloc not updating the file and line when using
recycled units.  Moral of the story: debugging code can shoot you in
the foot as well as help you.  And that wraps up 36 hours in the life
of a head banger.