Mainly splitting the sample rate converter stuff up into separate files.  It was just too big and ugly to work with otherwise.
Can someone update the project files for windows and check that
convert_acm compiles?  Can you also select windows converter, use it
for a few seconds, then switch back to 8kHz mono just to check that
destroying converter does not cause problems.  I had to hack the code
a bit to get it in a more separable form and may have broken
something.  Basically, if it can be created and destroyed it will work
fine.  Sorry to burden.

Fixed glitch in repair code I think.  Before repair was invoked if
next playout unit did not match expected time, not it is if next
playout is greater than or equal too.  Sometimes it was less than and
triggered repair.  Not entirely sure how this occurs.