If <keypress-q> on Entry don't quit. else quit :-)
I was frigging some pictures and tried to enter a name with a q in and
this was not a good move :-) Took longer than hoped to fix.  What
stumped me was that tk reports the parent widget in %W if the event
type is not relevent to the widget. Thus:

	bind .a.button <q> {puts "%W"}
	bind .a.entry  <q> {puts "%W"}

when q is pressed over .a.button the result is .
when q is pressed over .a.entry  the result is .a.entry

i guess it makes sense since the event gets passed up the event tree
until it can be handled and toplevel can handle all events.  just
stumped me for a few mins...