Replaced divisions in interpolation converter that were not powers of two with a multiplication scaling factor and a power of 2 denominator.  Improves performance 4-6 fold for upsampling by 3,5,6.
i.e. have sequence that looked like:
	x[i] = 2 * last + 1 * new; x[i] /= 3;
	x[i] = 1 * last + 2 * new; x[i] /= 3;
	x[i] = new;

since 1/3 approximates to 5 / 16, now looks like:

	x[i] = 10 * last +  5 * new; x[i] /= 16;
	x[i] =  5 * last + 10 * new; x[i] /= 16;
	x[i] =             15 * new; x[i] /= 16;

Have to scale last entry so not put jumps in amplitude.  I vaguely
recall reading that power of 2 div was quicker the rest.  Took a good
while to trace it though since my initial guesses were arithmetic
pipelining and memory alignment ;-)

Not sure we really care about this since the time is trivial anyway,
but gave odd data points in performance measurement for thesis.  For
the record, an UltraSPARC-II 296 MHz can upsample 330000 20ms frames
from 8k to 48k in 1 second.  Okay, the sound quality is not great, but
that's still a staggering amount of maths.