- Add check for newpcm/luigipcm and modify behaviour accordingly.
NB loopback mixer setting is broken under FreeBSD-4.0.  Appears to be
no way of handling this with newpcm at the moment (new ioctls are not
implemented in latest snapshot, 4.0-20000307-CURRENT).  When newpcm is
fixed we obviously have to go back and modify the driver interface to
support luigi style and newpcm style mixer.