Fix for gcc 3.3x as installed as a secondary compiler.  We end up needing snprintf and we clearly intended to use vsprintf in the replacement case but used sprintf.
This means we end up printing an address value in the controller
address string, which ends up being greater that 2x30-1, which in turn
breaks the parsing in the engine of the parent PID since it uses atoi
rather than strtoul.

The UI uses strtoul but nevers starts up since the messages it
receives from the engine have the wrong id in.

Five minute fix for FreeBSD 5.2/current port :-) Ho-Ho-Ho-Ho.