Two changes added: 1) New "security icon" (a padlock ala firefox) - which indicates when encryption is activated. This is activated when an mbus message is recieved, or thru the UI. Also fixed the key handling in rx_security_encryption_key() which wasn't setting the key in the session_t *sp before calling the ui_send_encryption_key() which was leading to spurious behaviour when the user changed the encryption state in the UI.
2) Modified rx_rtp_addr() so that it can be used to dynamically change the
address/port/ttl that RAT listens to - this is not possible thru a modified UI - 
which isn't unlike a web browser style. Some of the code is little nasty but it seems
to do the job - it should also be useful for AccessGrid to be able change session
address without having to restart RAT every time. Also in the next release RAT will
just start when 'double clicked' - like most applications these days and user can 
type in a new address once its started.