RAT now issues an error dialogue and quits when there are problems on startup (mbus rendezvous).
RAT now uses a timeout on [updated] rendezvous calls when RAT starts up and attempts rendezvous  between ratmedia and ratui. If the timeout is exceeded it is assumed that there is a problem - it usually indicates firewall/VPN problems. This also means that non-functioning ratmedia/ratui processes don't hang about after a failed startup. However it doesn't guarantee that rat processes won't hang around if other error conditions occur but this does mean the user is more aware of problems caused at startup - as opposed to being met with a silent exit.